+++ ibtk/include/ibtk/private/LSet-inl.h
@@ -215,7 +215,7 @@ inline size_t LSet<T>::getDataStreamSize() const
 template <class T>
 inline void LSet<T>::packStream(SAMRAI::tbox::AbstractStream& stream)
 {
-    int num_idx = d_set.size();
+    int num_idx = static_cast<int>(d_set.size());
     stream.pack(&num_idx, 1);
     for (unsigned int k = 0; k < d_set.size(); ++k)
     {
@@ -243,7 +243,7 @@ inline void LSet<T>::unpackStream(SAMRAI::tbox::AbstractStream& stream,
 template <class T>
 inline void LSet<T>::putToDatabase(SAMRAI::tbox::Pointer<SAMRAI::tbox::Database> database)
 {
-    const size_t data_sz = getDataStreamSize();
+    const int data_sz = static_cast<int>(getDataStreamSize());
     FixedSizedStream stream(data_sz);
     packStream(stream);
     database->putInteger("data_sz", data_sz);
@@ -256,7 +256,7 @@ template <class T>
 inline void LSet<T>::getFromDatabase(SAMRAI::tbox::Pointer<SAMRAI::tbox::Database> database)
 {
     database->getIntegerArray("d_offset", d_offset, NDIM);
-    const size_t data_sz = database->getInteger("data_sz");
+    const int data_sz = database->getInteger("data_sz");
     std::vector<char> data(data_sz);
     database->getCharArray("data", &data[0], data_sz);
     FixedSizedStream stream(&data[0], data_sz);
